requirements.txt
uvicorn[standard]>=0.30.6,<1
azure-data-tables>=12.5.0,<13
typer>=0.12.4,<1
# TODO: fix to a version once we are stable
watcloud-utils @ git+https://github.com/WATonomous/watcloud-utils.git
apscheduler>=3.10.4,<4

